rename package name to jsonencoder
