+++ sdk/schemaregistry/azure-schemaregistry-jsonencoder/LICENSE
