+++ sdk/schemaregistry/azure-schemaregistry-jsonencoder/assets.json
