+++ sdk/schemaregistry/azure-schemaregistry-jsonencoder/azure/schemaregistry/encoder/jsonencoder/_constants.py
