+++ sdk/schemaregistry/azure-schemaregistry-jsonencoder/azure/schemaregistry/encoder/jsonencoder/_exceptions.py
