+++ sdk/schemaregistry/azure-schemaregistry-jsonencoder/azure/schemaregistry/encoder/jsonencoder/_message_protocol.py
