+++ sdk/schemaregistry/azure-schemaregistry-jsonencoder/mypy.ini
