+++ sdk/schemaregistry/azure-schemaregistry-jsonencoder/samples/async_samples/encode_and_decode_event_data_message_async.py
@@ -24,16 +24,16 @@
 #
 # --------------------------------------------------------------------------
 """
-FILE: encode_and_decode_with_message_content_async.py
+FILE: encode_and_decode_event_data_message_async.py
 DESCRIPTION:
     This sample demonstrates the following:
      - Authenticating an async SchemaRegistryClient to be used by the JsonSchemaEncoder.
-     - Passing in content and schema to the JsonSchemaEncoder, which will return a dict containing
-      encoded content and corresponding content type.
-     - Passing in a dict containing Json-encoded content and corresponding content type to
-      the JsonSchemaEncoder, which will return the decoded content.
+     - Passing in content, schema, and EventData class to the JsonSchemaEncoder, which will return an
+      EventData object containing encoded content and corresponding content type.
+     - Passing in an `EventData` object with `body` set to Json-encoded content and `content_type`
+      set to corresponding content type to the JsonSchemaEncoder, which will return the decoded content.
 USAGE:
-    python encode_and_decode_with_message_content_async.py
+    python encode_and_decode_event_data_message_async.py
     Set the environment variables with your own values before running the sample:
     1) AZURE_TENANT_ID - The ID of the service principal's tenant. Also called its 'directory' ID.
     2) AZURE_CLIENT_ID - The service principal's client ID. Also called its 'application' ID.
@@ -52,13 +52,12 @@ import json
 
 from azure.identity.aio import ClientSecretCredential
 from azure.schemaregistry.aio import SchemaRegistryClient
-from azure.schemaregistry.encoder.jsonschemaencoder import MessageContent
-from azure.schemaregistry.encoder.jsonschemaencoder.aio import JsonSchemaEncoder
+from azure.schemaregistry.encoder.jsonencoder.aio import JsonSchemaEncoder
 from azure.eventhub import EventData
 
-TENANT_ID = os.environ["AZURE_TENANT_ID"]
-CLIENT_ID = os.environ["AZURE_CLIENT_ID"]
-CLIENT_SECRET = os.environ["AZURE_CLIENT_SECRET"]
+TENANT_ID=os.environ['AZURE_TENANT_ID']
+CLIENT_ID=os.environ['AZURE_CLIENT_ID']
+CLIENT_SECRET=os.environ['AZURE_CLIENT_SECRET']
 
 SCHEMAREGISTRY_FULLY_QUALIFIED_NAMESPACE = os.environ['SCHEMAREGISTRY_JSON_FULLY_QUALIFIED_NAMESPACE']
 GROUP_NAME = os.environ['SCHEMAREGISTRY_GROUP']
@@ -87,28 +86,39 @@ SCHEMA_STRING = json.dumps(SCHEMA_JSON)
 
 
 token_credential = ClientSecretCredential(
-    tenant_id=TENANT_ID, client_id=CLIENT_ID, client_secret=CLIENT_SECRET
+    tenant_id=TENANT_ID,
+    client_id=CLIENT_ID,
+    client_secret=CLIENT_SECRET
 )
 
 
-async def encode_message_content_dict(encoder):
+async def encode_to_event_data_message(encoder):
     dict_content_ben = {"name": "Ben", "favorite_number": 7, "favorite_color": "red"}
-    encoded_message_content_ben = await encoder.encode(dict_content_ben, schema=SCHEMA_STRING)
+    dict_content_alice = {"name": "Alice", "favorite_number": 15, "favorite_color": "green"}
 
-    print("Encoded message_content is: ", encoded_message_content_ben)
-    return EventData.from_message_content(
-        encoded_message_content_ben["content"],
-        encoded_message_content_ben["content_type"],
+    # Schema would be automatically registered into Schema Registry and cached locally.
+    event_data_ben = await encoder.encode(
+        dict_content_ben, schema=SCHEMA_STRING, message_type=EventData
     )
 
-async def decode_with_content_and_content_type(encoder, event_data):
-    # get content as bytes
-    content = bytearray()
-    for d in event_data.body:
-        content += d
-    content_bytes = bytes(content)
-    message_content = MessageContent({"content": content_bytes, "content_type": event_data.content_type})
-    decoded_content = await encoder.decode(message_content)
+    # The second call won't trigger a service call.
+    event_data_alice = await encoder.encode(
+        dict_content_alice, schema=SCHEMA_STRING, message_type=EventData
+    )
+
+    print("Encoded content is: ", next(event_data_ben.body))
+    print("Encoded content is: ", next(event_data_alice.body))
+
+    print("Encoded content type is: ", event_data_ben.content_type)
+    print("Encoded content type is: ", event_data_alice.content_type)
+    return [event_data_ben, event_data_alice]
+
+
+async def decode_event_data_message(encoder, event_data):
+    # encoder.decode would extract the schema id from the content_type,
+    # retrieve schema from Schema Registry and cache the schema locally.
+    # If the schema id is in the local cache, the call won't trigger a service call.
+    decoded_content = await encoder.decode(event_data)
 
     print("Decoded content is: ", decoded_content)
     return decoded_content
@@ -122,8 +132,9 @@ async def main():
     encoder = JsonSchemaEncoder(
         client=schema_registry, group_name=GROUP_NAME, auto_register=True
     )
-    event_data = await encode_message_content_dict(encoder)
-    decoded_content = await decode_with_content_and_content_type(encoder, event_data)
+    event_data_ben, event_data_alice = await encode_to_event_data_message(encoder)
+    decoded_content_ben = await decode_event_data_message(encoder, event_data_ben)
+    decoded_content_alice = await decode_event_data_message(encoder, event_data_alice)
     await encoder.close()
     await token_credential.close()
