+++ sdk/schemaregistry/azure-schemaregistry-jsonencoder/sdk_packaging.toml
