+++ sdk/schemaregistry/azure-schemaregistry-jsonencoder/tests/perfstress_tests/decode_content.py
