+++ sdk/schemaregistry/azure-schemaregistry-jsonencoder/tests/perfstress_tests/encode_content.py
