+++ sdk/schemaregistry/azure-schemaregistry-jsonencoder/tests/test_json_encoder.py
@@ -26,23 +26,22 @@
 import functools
 import pytest
 import json
-import pytest
 import jsonschema
 from genson import SchemaBuilder
 from typing import Mapping, Any
 
-from azure.schemaregistry import SchemaFormat
-from azure.schemaregistry.aio import SchemaRegistryClient
-from azure.schemaregistry.encoder.jsonschemaencoder.aio import JsonSchemaEncoder
-from azure.schemaregistry.encoder.jsonschemaencoder import InvalidSchemaError, InvalidContentError
-
-from devtools_testutils import AzureRecordedTestCase, EnvironmentVariableLoader
-from devtools_testutils.aio import recorded_by_proxy_async
+from azure.schemaregistry import SchemaRegistryClient, SchemaFormat
+from azure.schemaregistry.encoder.jsonencoder import JsonSchemaEncoder, InvalidContentError, InvalidSchemaError
+from devtools_testutils import (
+    AzureRecordedTestCase,
+    EnvironmentVariableLoader,
+    recorded_by_proxy,
+)
 
 SchemaRegistryEnvironmentVariableLoader = functools.partial(
     EnvironmentVariableLoader,
     "schemaregistry",
-    schemaregistry_json_fully_qualified_namespace="fake_resource_json.servicebus.windows.net",
+    schemaregistry_json_fully_qualified_namespace="fake_resource_json.servicebus.windows.net/",
     schemaregistry_group="fakegroup"
 )
 
@@ -69,27 +68,28 @@ SCHEMA_JSON = {
 SCHEMA_STR = json.dumps(SCHEMA_JSON)
 SCHEMA_BYTES = SCHEMA_STR.encode()
 
-class TestJsonSchemaEncoderAsync(AzureRecordedTestCase):
+class TestJsonSchemaEncoder(AzureRecordedTestCase):
 
-    def create_client(self, **kwargs):
+    def create_client(self, **kwargs) -> SchemaRegistryClient:
         fully_qualified_namespace = kwargs.pop("fully_qualified_namespace")
-        credential = self.get_credential(SchemaRegistryClient, is_async=True)
-        return self.create_client_from_credential(SchemaRegistryClient, credential, fully_qualified_namespace=fully_qualified_namespace, is_async=True)
+        credential = self.get_credential(SchemaRegistryClient)
+        return self.create_client_from_credential(SchemaRegistryClient, credential, fully_qualified_namespace=fully_qualified_namespace)
 
-    @pytest.mark.asyncio
     @SchemaRegistryEnvironmentVariableLoader()
-    @recorded_by_proxy_async
-    async def test_sr_json_encoder_encode_decode_with_str_schema(self, schemaregistry_json_fully_qualified_namespace, schemaregistry_group, **kwargs):
+    @recorded_by_proxy
+    def test_sr_json_encoder_encode_decode_with_str_schema(self, **kwargs):
+        schemaregistry_json_fully_qualified_namespace = kwargs.pop("schemaregistry_json_fully_qualified_namespace")
+        schemaregistry_group = kwargs.pop("schemaregistry_group")
         sr_client = self.create_client(fully_qualified_namespace=schemaregistry_json_fully_qualified_namespace)
         sr_json_encoder = JsonSchemaEncoder(client=sr_client, group_name=schemaregistry_group)
 
         # register schema first
-        schema_properties = await sr_client.register_schema(schemaregistry_group, SCHEMA_JSON["title"], SCHEMA_STR, SchemaFormat.JSON)
+        schema_properties = sr_client.register_schema(schemaregistry_group, SCHEMA_JSON["title"], SCHEMA_STR, SchemaFormat.JSON)
         schema_id = schema_properties.id
 
         # encode w/ str schema
         dict_content = {"name": u"Ben", "favorite_number": 7, "favorite_color": u"red"}
-        encoded_message_content = await sr_json_encoder.encode(dict_content, schema=SCHEMA_STR)
+        encoded_message_content = sr_json_encoder.encode(dict_content, schema=SCHEMA_STR)
         content_type = encoded_message_content["content_type"]
         encoded_content = encoded_message_content["content"]
         assert content_type.split("+")[0] == 'application/json'
@@ -97,7 +97,7 @@ class TestJsonSchemaEncoderAsync(AzureRecordedTestCase):
 
         # decode
         encoded_content_dict = {"content": encoded_content, "content_type": content_type}
-        decoded_content = await sr_json_encoder.decode(encoded_content_dict)
+        decoded_content = sr_json_encoder.decode(encoded_content_dict)
         assert decoded_content["name"] == u"Ben"
         assert decoded_content["favorite_number"] == 7
         assert decoded_content["favorite_color"] == u"red"
@@ -106,11 +106,11 @@ class TestJsonSchemaEncoderAsync(AzureRecordedTestCase):
         mime_type, schema_id = encoded_content_dict["content_type"].split("+")
         encoded_content_dict["content_type"] = "binary/fake+" + schema_id
         with pytest.raises(InvalidContentError) as e:
-            decoded_content = await sr_json_encoder.decode(encoded_content_dict)
+            decoded_content = sr_json_encoder.decode(encoded_content_dict)
 
         encoded_content_dict["content_type"] = 'a+b+c'
         with pytest.raises(InvalidContentError) as e:
-            decoded_content = await sr_json_encoder.decode(encoded_content_dict)
+            decoded_content = sr_json_encoder.decode(encoded_content_dict)
 
         # check that JsonSchemaEncoder won't work with message types that don't follow protocols
         class BadExample:
@@ -118,12 +118,12 @@ class TestJsonSchemaEncoderAsync(AzureRecordedTestCase):
                 self.not_content = not_content
 
         with pytest.raises(TypeError) as e:
-            await sr_json_encoder.encode({"name": u"Ben"}, schema=SCHEMA_STR, message_type=BadExample) 
+            sr_json_encoder.encode({"name": u"Ben"}, schema=SCHEMA_STR, message_type=BadExample) 
         assert "subtype of the MessageType" in (str(e.value))
 
         bad_ex = BadExample('fake')
         with pytest.raises(TypeError) as e:    # caught TypeError
-           await sr_json_encoder.decode(message=bad_ex) 
+            sr_json_encoder.decode(message=bad_ex) 
         assert "subtype of the MessageType" in (str(e.value))
 
         # check that JsonSchemaEncoder will work with message types that follow protocols
@@ -143,42 +143,43 @@ class TestJsonSchemaEncoderAsync(AzureRecordedTestCase):
                 return {"content": self.content, "content_type": self.content_type}
 
         # test that extra kwargs pass through to message_type constructor
-        good_ex_obj = await sr_json_encoder.encode(dict_content, schema=SCHEMA_STR, message_type=GoodExample, extra='val') 
-        decoded_content_obj = await sr_json_encoder.decode(message=good_ex_obj)
+        good_ex_obj = sr_json_encoder.encode(dict_content, schema=SCHEMA_STR, message_type=GoodExample, extra='val') 
+        decoded_content_obj = sr_json_encoder.decode(message=good_ex_obj)
 
         assert decoded_content_obj["name"] == u"Ben"
         assert decoded_content_obj["favorite_number"] == 7
         assert decoded_content_obj["favorite_color"] == u"red"
-        await sr_client.close()
-        await sr_json_encoder.close()
+        sr_client.close()
+        sr_json_encoder.close()
 
         # no group_name passed into constructor, check encode fails, but decode works
         extra_sr_client = self.create_client(fully_qualified_namespace=schemaregistry_json_fully_qualified_namespace)
         sr_json_encoder_no_group = JsonSchemaEncoder(client=extra_sr_client, auto_register=True)
-        decoded_content = await sr_json_encoder_no_group.decode(encoded_message_content)
+        decoded_content = sr_json_encoder_no_group.decode(encoded_message_content)
         assert decoded_content["name"] == u"Ben"
         assert decoded_content["favorite_number"] == 7
         assert decoded_content["favorite_color"] == u"red"
         with pytest.raises(TypeError):
-            encoded_message_content = await sr_json_encoder_no_group.encode(dict_content, schema=SCHEMA_STR)
-        await sr_json_encoder_no_group.close()
-        await extra_sr_client.close()
+            encoded_message_content = sr_json_encoder_no_group.encode(dict_content, schema=SCHEMA_STR)
+        sr_json_encoder_no_group.close()
+        extra_sr_client.close()
 
-    @pytest.mark.asyncio
     @SchemaRegistryEnvironmentVariableLoader()
-    @recorded_by_proxy_async
-    async def test_sr_json_encoder_encode_decode_with_bytes_schema(self, schemaregistry_json_fully_qualified_namespace, schemaregistry_group, **kwargs):
+    @recorded_by_proxy
+    def test_sr_json_encoder_encode_decode_with_bytes_schema(self, **kwargs):
+        schemaregistry_json_fully_qualified_namespace = kwargs.pop("schemaregistry_json_fully_qualified_namespace")
+        schemaregistry_group = kwargs.pop("schemaregistry_group")
         sr_client = self.create_client(fully_qualified_namespace=schemaregistry_json_fully_qualified_namespace)
         # pass default schema
         sr_json_encoder = JsonSchemaEncoder(client=sr_client, group_name=schemaregistry_group, schema=SCHEMA_BYTES)
 
         # register schema first
-        schema_properties = await sr_client.register_schema(schemaregistry_group, SCHEMA_JSON["title"], SCHEMA_STR, SchemaFormat.JSON)
+        schema_properties = sr_client.register_schema(schemaregistry_group, SCHEMA_JSON["title"], SCHEMA_STR, SchemaFormat.JSON)
         schema_id = schema_properties.id
 
         # encode w/ bytes schema
         dict_content = {"name": u"Ben", "favorite_number": 7, "favorite_color": u"red"}
-        encoded_message_content = await sr_json_encoder.encode(dict_content)
+        encoded_message_content = sr_json_encoder.encode(dict_content)
         content_type = encoded_message_content["content_type"]
         encoded_content = encoded_message_content["content"]
         assert content_type.split("+")[0] == 'application/json'
@@ -186,16 +187,17 @@ class TestJsonSchemaEncoderAsync(AzureRecordedTestCase):
 
         # decode
         encoded_content_dict = {"content": encoded_content, "content_type": content_type}
-        decoded_content = await sr_json_encoder.decode(encoded_content_dict)
+        decoded_content = sr_json_encoder.decode(encoded_content_dict)
         assert decoded_content["name"] == u"Ben"
         assert decoded_content["favorite_number"] == 7
         assert decoded_content["favorite_color"] == u"red"
-        await sr_json_encoder.close()
+        sr_json_encoder.close()
 
-    @pytest.mark.asyncio
     @SchemaRegistryEnvironmentVariableLoader()
-    @recorded_by_proxy_async
-    async def test_sr_json_encoder_encode_decode_with_generate_schema_callable(self, schemaregistry_json_fully_qualified_namespace, schemaregistry_group, **kwargs):
+    @recorded_by_proxy
+    def test_sr_json_encoder_encode_decode_with_generate_schema_callable(self, **kwargs):
+        schemaregistry_json_fully_qualified_namespace = kwargs.pop("schemaregistry_json_fully_qualified_namespace")
+        schemaregistry_group = kwargs.pop("schemaregistry_group")
         sr_client = self.create_client(fully_qualified_namespace=schemaregistry_json_fully_qualified_namespace)
         sr_json_encoder = JsonSchemaEncoder(client=sr_client, group_name=schemaregistry_group, schema=SCHEMA_BYTES)
 
@@ -211,7 +213,7 @@ class TestJsonSchemaEncoderAsync(AzureRecordedTestCase):
 
         # encode w/ callable + override default schema
         dict_content = {"name": u"Ben", "favorite_number": 7, "favorite_color": u"red"}
-        encoded_message_content = await sr_json_encoder.encode(dict_content, schema=generate_schema)
+        encoded_message_content = sr_json_encoder.encode(dict_content, schema=generate_schema)
         content_type = encoded_message_content["content_type"]
         encoded_content = encoded_message_content["content"]
         assert content_type.split("+")[0] == 'application/json'
@@ -222,20 +224,21 @@ class TestJsonSchemaEncoderAsync(AzureRecordedTestCase):
             pass
 
         with pytest.raises(InvalidSchemaError):
-            encoded_message_content = await sr_json_encoder.encode(invalid_content, schema=generate_schema)
+            encoded_message_content = sr_json_encoder.encode(invalid_content, schema=generate_schema)
 
         # decode
         encoded_content_dict = {"content": encoded_content, "content_type": content_type}
-        decoded_content = await sr_json_encoder.decode(encoded_content_dict)
+        decoded_content = sr_json_encoder.decode(encoded_content_dict)
         assert decoded_content["name"] == u"Ben"
         assert decoded_content["favorite_number"] == 7
         assert decoded_content["favorite_color"] == u"red"
-        await sr_json_encoder.close()
+        sr_json_encoder.close()
 
-    @pytest.mark.asyncio
     @SchemaRegistryEnvironmentVariableLoader()
-    @recorded_by_proxy_async
-    async def test_sr_json_encoder_validate_content(self, schemaregistry_json_fully_qualified_namespace, schemaregistry_group, **kwargs):
+    @recorded_by_proxy
+    def test_sr_json_encoder_validate_content(self, **kwargs):
+        schemaregistry_json_fully_qualified_namespace = kwargs.pop("schemaregistry_json_fully_qualified_namespace")
+        schemaregistry_group = kwargs.pop("schemaregistry_group")
         sr_client = self.create_client(fully_qualified_namespace=schemaregistry_json_fully_qualified_namespace)
 
         def default_validate(content: Mapping[str, Any], schema: Mapping[str, Any]) -> None:
@@ -245,7 +248,7 @@ class TestJsonSchemaEncoderAsync(AzureRecordedTestCase):
 
         # encode w/ validation
         dict_content = {"name": u"Ben", "favorite_number": 7, "favorite_color": u"red"}
-        encoded_message_content = await sr_json_encoder.encode(dict_content, schema=SCHEMA_STR)
+        encoded_message_content = sr_json_encoder.encode(dict_content, schema=SCHEMA_STR)
         content_type = encoded_message_content["content_type"]
         encoded_content = encoded_message_content["content"]
         assert content_type.split("+")[0] == 'application/json'
@@ -253,7 +256,7 @@ class TestJsonSchemaEncoderAsync(AzureRecordedTestCase):
 
         # decode w/ validation
         encoded_content_dict = {"content": encoded_content, "content_type": content_type}
-        decoded_content = await sr_json_encoder.decode(encoded_content_dict)
+        decoded_content = sr_json_encoder.decode(encoded_content_dict)
         assert decoded_content["name"] == u"Ben"
         assert decoded_content["favorite_number"] == 7
         assert decoded_content["favorite_color"] == u"red"
@@ -261,10 +264,10 @@ class TestJsonSchemaEncoderAsync(AzureRecordedTestCase):
         # wrong data type, check with validation
         dict_content_bad = {"name": u"Ben", "favorite_number": 7, "favorite_color": 7}
         with pytest.raises(InvalidContentError) as e:
-            encoded_message_content = await sr_json_encoder.encode(dict_content_bad, schema=SCHEMA_STR)
+            encoded_message_content = sr_json_encoder.encode(dict_content_bad, schema=SCHEMA_STR)
         assert "schema_id" in e.value.details
 
         # override validation, bad content isn't detected
         dict_content_bad = {"name": u"Ben", "favorite_number": 7, "favorite_color": 7}
-        encoded_message_content = await sr_json_encoder.encode(dict_content_bad, schema=SCHEMA_STR, validate=False)
-        await sr_json_encoder.close()
+        encoded_message_content = sr_json_encoder.encode(dict_content_bad, schema=SCHEMA_STR, validate=False)
+        sr_json_encoder.close()
